+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/LnurlAuthFlowTestBase.java
@@ -1,14 +1,7 @@
 package org.tbk.spring.lnurl.security;
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.MockMvcPrint;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.mock.web.MockHttpSession;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.tbk.lnurl.auth.LinkingKey;
@@ -17,49 +10,27 @@ import org.tbk.lnurl.auth.SignedLnurlAuth;
 import org.tbk.lnurl.simple.SimpleLnurl;
 import org.tbk.lnurl.simple.auth.SimpleLinkingKey;
 import org.tbk.lnurl.simple.auth.SimpleLnurlAuth;
-import org.tbk.lnurl.test.SimpleLnurlWallet;
-import org.tbk.spring.lnurl.security.test.app.LnurlAuthTestApplication;
+import org.tbk.lnurl.test.LnurlWallet;
 
-import java.security.SecureRandom;
 import java.util.Optional;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@SpringBootTest(
-        webEnvironment = WebEnvironment.RANDOM_PORT,
-        classes = LnurlAuthTestApplication.class
-)
-@AutoConfigureMockMvc(print = MockMvcPrint.LOG_DEBUG, printOnlyOnFailure = false)
-@ActiveProfiles("test")
-class LnurlAuthFlowTest {
-    private static final SecureRandom random = new SecureRandom();
+class LnurlAuthFlowTestBase {
 
-    @Autowired
-    private MockMvc mockMvc;
-
-    private static SimpleLnurlWallet testWallet;
-
-    @BeforeAll
-    static void setUpAll() {
-        byte[] seed = random.generateSeed(256);
-        testWallet = SimpleLnurlWallet.fromSeed(seed);
-    }
-
-    @Test
-    void lnurlAuthLoginSuccess() throws Exception {
+    static void testLnurlAuthLoginSuccess(MockMvc mockMvc, LnurlWallet testWallet) throws Exception {
         /*
          * STEP 1: Create a session for web user (containing a newly created 'k1' value)
          *
-         * Create a session with k1 value and return an bech32 encoded lnurl-auth string.
+         * Create a session with k1 value and return a bech32 encoded lnurl-auth string.
          * Only the browser knowing the session ID can login after a wallet signed the k1 value.
          *
          * This is done in a browser.
-         * A login page can, e.g. display an qr-code for wallets to scan.
+         * A login page can, e.g. display a qr-code for wallets to scan.
          */
         MvcResult loginResult = mockMvc.perform(get("/login"))
                 .andReturn();
@@ -86,7 +57,7 @@ class LnurlAuthFlowTest {
 
         // assert that the user still cannot see any guarded resource
         mockMvc.perform(get("/authenticated")
-                .sessionAttr(LnurlAuthConfigurer.defaultSessionK1Key(), lnurlAuth.getK1().toHex()))
+                        .sessionAttr(LnurlAuthConfigurer.defaultSessionK1Key(), lnurlAuth.getK1().toHex()))
                 .andExpect(status().isForbidden());
 
         /*
@@ -103,7 +74,7 @@ class LnurlAuthFlowTest {
          * - etc.
          */
         MvcResult migrateResult = mockMvc.perform(get(LnurlAuthConfigurer.defaultSessionLoginUrl())
-                .sessionAttr(LnurlAuthConfigurer.defaultSessionK1Key(), lnurlAuth.getK1().toHex()))
+                        .sessionAttr(LnurlAuthConfigurer.defaultSessionK1Key(), lnurlAuth.getK1().toHex()))
                 .andExpect(status().is3xxRedirection())
                 .andExpect(header().string(HttpHeaders.LOCATION, is("/")))
                 .andExpect(redirectedUrl("/"))
@@ -117,11 +88,11 @@ class LnurlAuthFlowTest {
         assertThat("Session has been migrated", afterLoginCompletedSession.getId(), is(not(beforeLoginCompletedSession.getId())));
 
         /*
-         * STEP 4: User is now logged in an can access guarded resources.
+         * STEP 4: User is now logged in and can access guarded resources.
          */
-        // authenticated has been setup to require role "LNURL_AUTH_TEST_USER"
+        // authenticated has been configured to require role "LNURL_AUTH_TEST_USER"
         MvcResult authenticatedResult = mockMvc.perform(get("/authenticated")
-                .session(afterLoginCompletedSession))
+                        .session(afterLoginCompletedSession))
                 .andExpect(status().isOk())
                 .andReturn();