+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/test/app1/LnurlAuthTestApplication.java
@@ -1,10 +1,13 @@
-package org.tbk.spring.lnurl.security.test.app;
+package org.tbk.spring.lnurl.security.test.app1;
 
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.Import;
+import org.tbk.spring.lnurl.security.test.shared.LnurlAuthTestConfig;
 
 @SpringBootApplication
+@Import(LnurlAuthTestConfig.class)
 public class LnurlAuthTestApplication {
 
     public static void main(String[] args) {