+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/test/app2/LnurlAuthTestApplicationWithDeprecatedConfig.java
@@ -1,0 +1,20 @@
+package org.tbk.spring.lnurl.security.test.app2;
+
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.Import;
+import org.tbk.spring.lnurl.security.test.shared.LnurlAuthTestConfig;
+
+@SpringBootApplication
+@Import(LnurlAuthTestConfig.class)
+public class LnurlAuthTestApplicationWithDeprecatedConfig {
+
+    public static void main(String[] args) {
+        new SpringApplicationBuilder()
+                .sources(LnurlAuthTestApplicationWithDeprecatedConfig.class)
+                .web(WebApplicationType.SERVLET)
+                .profiles("test")
+                .run(args);
+    }
+}