+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/test/app2/LnurlAuthTestSecurityDeprecatedConfig.java
@@ -1,27 +1,33 @@
-package org.tbk.spring.lnurl.security.test.app;
+package org.tbk.spring.lnurl.security.test.app2;
 
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.firewall.StrictHttpFirewall;
-import org.tbk.lnurl.auth.LnurlAuthFactory;
-import org.tbk.lnurl.auth.SimpleK1Manager;
-import org.tbk.lnurl.auth.SimpleLnurlAuthFactory;
+import org.tbk.lnurl.auth.*;
 import org.tbk.spring.lnurl.security.LnurlAuthConfigurer;
-
-import java.net.URI;
 
 @Slf4j
 @EnableWebSecurity
 @Configuration
-public class LnurlAuthTestSecurityConfig extends WebSecurityConfigurerAdapter {
+@RequiredArgsConstructor
+public class LnurlAuthTestSecurityDeprecatedConfig extends WebSecurityConfigurerAdapter {
+
+    @NonNull
+    private final LnurlAuthPairingService pairingService;
+
+    @NonNull
+    private final K1Manager k1Manager;
+    @NonNull
+    private final UserDetailsService userDetailsService;
 
     @Override
     public void configure(WebSecurity web) {
@@ -31,8 +37,8 @@ public class LnurlAuthTestSecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http
-                .userDetailsService(userDetailsService())
-                .apply(LnurlAuthConfigurer.create(k1Manager(), pairingService()))
+                .userDetailsService(userDetailsService)
+                .apply(LnurlAuthConfigurer.create(k1Manager, pairingService))
                 .and()
                 .sessionManagement(session -> session
                         .sessionCreationPolicy(SessionCreationPolicy.NEVER)
@@ -46,29 +52,4 @@ public class LnurlAuthTestSecurityConfig extends WebSecurityConfigurerAdapter {
                         .anyRequest().authenticated()
                 );
     }
-
-    @Bean
-    public LnurlAuthFactory lnurlAuthFactory() {
-        // any url well do - we just need the query string in tests
-        URI loginUrl = URI.create("https://localhost" + LnurlAuthConfigurer.defaultWalletLoginUrl());
-
-        return new SimpleLnurlAuthFactory(loginUrl, k1Manager());
-    }
-
-    @Bean
-    public TestPairingService pairingService() {
-        return new TestPairingService(userDetailsService());
-    }
-
-    @Bean
-    public InMemoryUserDetailsManager userDetailsService() {
-        return new InMemoryUserDetailsManager();
-    }
-
-    // make bean of K1Manager injectable in tests
-    @Bean
-    public SimpleK1Manager k1Manager() {
-        return new SimpleK1Manager();
-    }
-
 }